precompiles: Fix BN254 input point validation
Add missing check for the prime field elements being out of range
when loading from bytes an input point to any BN254 precompile.

This bug was hidden because the out-of-range values were accidentally
reduced with "mod P". So to reveal the bug we needed a special input
point such that (x mod P, y mod P) passed the curve equation.

The bug was introduced in the initial implementationof the BN254
precompiles: https://github.com/ipsilon/evmone/pull/716